Solving the driver not found issue
Overlook this error during development. I'm hard coding the valid ODBC
drivers but not taking into consideration the installed ODBC drivers on
other host machines! Lesson learned, always take into consideration the
platform your app will be deployed during development. If it works on
your machine doesn't mean that it will work for them. Just keep that in
mind.

Testing this latest solution tommorrow. If all is running smooth and
well, this will be the latest release for v0.2.