Applying last minute amendments!
This should not be happening but the additional brief descriptions added
in the constant.py is crucial! :smile:

I believe this is the last commit for the development of 0.3. Currently
preparing now for the release of this version. This has been another
journey of learning. The challenges are are truly worth it.

Happy Coding!